handle eldap extensibleMatch boolean bug
eldap has a bug in the dnAttributes boolean option of the extensibleMatch filter helper,
where a string is set as TRUE / FALSE istead of a bool.
This leads in an ASN1 error when trying to submit the request to the server.
This will be removed at some time, when the OTP team fixes the issue.

A PR has been opened here: https://github.com/erlang/otp/pull/5615